refactor: useInternalBalancesTransaction to modular hooks
