+++ packages/next-app/src/app/internalmanager/[network]/[operationKind]/all/withdrawAll.tsx
@@ -21,10 +21,8 @@ import { Toast } from "#/components/Toast";
 import { Form } from "#/components/ui/form";
 import WalletNotConnected from "#/components/WalletNotConnected";
 import { useInternalBalance } from "#/contexts/InternalManagerContext";
-import {
-  TransactionStatus,
-  useInternalBalancesTransaction,
-} from "#/hooks/useTransaction";
+import { useManageUserBalance } from "#/hooks/useManageUserBalance";
+import { TransactionStatus } from "#/hooks/useTransaction";
 import { internalBalances } from "#/lib/gql";
 import { UserBalanceOpKind } from "#/lib/internal-balance-helper";
 import { refetchRequest } from "#/utils/fetcher";
@@ -65,10 +63,7 @@ export function WithdrawAll() {
 
   const { register, setValue } = form;
 
-  const { handleTransaction } = useInternalBalancesTransaction({
-    userAddress: addressLower as Address,
-    operationKind: UserBalanceOpKind.WITHDRAW_INTERNAL,
-  });
+  const { handleTransaction } = useManageUserBalance();
 
   function handleOnSubmit(data: FieldValues) {
     if (!tokenData?.user?.userInternalBalances) return;
@@ -80,7 +75,11 @@ export function WithdrawAll() {
         receiverAddress: data.receiverAddress,
       };
     });
-    handleTransaction({ data: tokenInfo });
+    handleTransaction({
+      data: tokenInfo,
+      operationKind: UserBalanceOpKind.WITHDRAW_INTERNAL,
+      userAddress: addressLower as Address,
+    });
   }
 
   useEffect(() => {
