+++ packages/next-app/src/app/internalmanager/[network]/[operationKind]/token/[tokenAddress]/page.tsx
@@ -31,10 +31,9 @@ import { Form } from "#/components/ui/form";
 import WalletNotConnected from "#/components/WalletNotConnected";
 import { useInternalBalance } from "#/contexts/InternalManagerContext";
 import { getNetwork } from "#/contexts/networks";
-import {
-  TransactionStatus,
-  useInternalBalancesTransaction,
-} from "#/hooks/useTransaction";
+import { useCheckAllowance } from "#/hooks/useCheckAllowance";
+import { useManageUserBalance } from "#/hooks/useManageUserBalance";
+import { TransactionStatus } from "#/hooks/useTransaction";
 import { internalBalances } from "#/lib/gql";
 import {
   operationKindType,
@@ -229,6 +228,9 @@ function TransactionCard({
       operationKindType[operationKindParam as keyof typeof operationKindType]
     ];
 
+  const { handleTransaction } = useManageUserBalance();
+  const { checkAllowance } = useCheckAllowance();
+
   const InternalBalanceSchema = getInternalBalanceSchema({
     totalBalance:
       operationKindEnum === UserBalanceOpKind.DEPOSIT_INTERNAL
@@ -243,10 +245,6 @@ function TransactionCard({
     resolver: zodResolver(InternalBalanceSchema),
   });
   const { register, setValue, watch } = form;
-  const { handleTransaction, checkAllowance } = useInternalBalancesTransaction({
-    userAddress: userAddress,
-    operationKind: operationKindEnum,
-  });
   const receiverAddressValue = watch("receiverAddress");
   const tokenAmount = watch("tokenAmount");
 
@@ -264,6 +262,7 @@ function TransactionCard({
         tokenAddress: tokenData.tokenInfo.address as Address,
         tokenAmount,
         tokenDecimals: tokenData.tokenInfo.decimals,
+        userAddress,
       });
     }
   }, [tokenAmount]);
@@ -280,7 +279,12 @@ function TransactionCard({
       tokenAmount: data.tokenAmount,
       tokenDecimals: tokenData.tokenInfo.decimals,
     };
-    handleTransaction({ data: [transactionData] });
+
+    handleTransaction({
+      data: [transactionData],
+      operationKind: operationKindEnum,
+      userAddress,
+    });
   }
 
   const { transactionStatus } = useInternalBalance();
